Set up NPM project structure
git-svn-id: https://svn.sosy-lab.org/software/cpachecker/branches/report-js-build@36305 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c
