+++ .gitignore
@@ -76,3 +76,5 @@ a.out
 .idea/findbugs-idea.xml
 .idea/junitgenerator-prj-settings.xml
 .idea/vcs.xml
+
+node_modules/
