+++ src/org/sosy_lab/cpachecker/core/counterexample/ReportGenerator.java
@@ -88,8 +88,12 @@ public class ReportGenerator {
   private static final Splitter LINE_SPLITTER = Splitter.on('\n');
 
   private static final String HTML_TEMPLATE = "report.html";
-  private static final String CSS_TEMPLATE = "report.css";
-  private static final String JS_TEMPLATE = "report.js";
+  private static final String CSS_TEMPLATE = "build/bundle.css";
+  private static final String JS_TEMPLATE = "build/bundle.js";
+  private static final String JS_ARGWORKER_TEMPLATE = "worker/argWorker.js";
+  private static final String JS_CFAWORKER_TEMPLATE = "worker/cfaWorker.js";
+  private static final String JS_D3_TEMPLATE = "external_libs/d3.min.js";
+  private static final String JS_DAGRE_D3_TEMPLATE = "external_libs/dagre-d3.min.js";
 
   private final Configuration config;
   private final LogManager logger;
@@ -256,6 +260,8 @@ public class ReportGenerator {
         } else if (line.contains("REPORT_CSS")) {
           insertCss(writer);
         } else if (line.contains("REPORT_JS")) {
+          insertWorkerJs(writer, JS_ARGWORKER_TEMPLATE);
+          insertWorkerJs(writer, JS_CFAWORKER_TEMPLATE);
           insertJs(writer, cfa, dotBuilder, counterExample);
         } else if (line.contains("STATISTICS")) {
           insertStatistics(writer, statistics);
@@ -288,24 +294,54 @@ public class ReportGenerator {
       DOTBuilder2 dotBuilder,
       @Nullable CounterexampleInfo counterExample)
       throws IOException {
-    try (BufferedReader reader =
-        Resources.asCharSource(Resources.getResource(getClass(), JS_TEMPLATE), Charsets.UTF_8)
-            .openBufferedStream();) {
-      String line;
-      while (null != (line = reader.readLine())) {
-        if (line.contains("CFA_JSON_INPUT")) {
-          insertCfaJson(writer, cfa, dotBuilder, counterExample);
-        } else if (line.contains("ARG_JSON_INPUT")) {
-          insertArgJson(writer);
-        } else if (line.contains("SOURCE_FILES")) {
-          insertSourceFileNames(writer);
-        } else {
-          writer.write(line);
-          writer.write('\n');
+        insertCfaJson(writer, cfa, dotBuilder, counterExample);
+        insertArgJson(writer);
+        insertSourceFileNames(writer);
+
+        try (BufferedReader reader =
+            Resources.asCharSource(Resources.getResource(getClass(), JS_TEMPLATE), Charsets.UTF_8)
+                .openBufferedStream();) {
+          String line;
+          while (null != (line = reader.readLine())) {
+              writer.write(line);
+              writer.write('\n');
+          }
         }
       }
-    }
-  }
+
+  private void insertWorkerJs(
+      Writer writer,
+      String jsFile)
+      throws IOException {
+        try (BufferedReader reader =
+            Resources.asCharSource(Resources.getResource(getClass(), jsFile), Charsets.UTF_8)
+                .openBufferedStream();) {
+          String line;
+          while (null != (line = reader.readLine())) {
+              if (line.contains("EXTERNAL_LIBS")) {
+                insertFile(writer, JS_D3_TEMPLATE);
+                insertFile(writer, JS_DAGRE_D3_TEMPLATE);
+              }
+              writer.write(line);
+              writer.write('\n');
+          }
+        }
+      }
+
+  private void insertFile(
+      Writer writer,
+      String file)
+      throws IOException {
+        try (BufferedReader reader =
+            Resources.asCharSource(Resources.getResource(getClass(), file), Charsets.UTF_8)
+                .openBufferedStream();) {
+          String line;
+          while (null != (line = reader.readLine())) {
+              writer.write(line);
+              writer.write('\n');
+          }
+        }
+      }
 
   private void insertCfaJson(
       Writer writer, CFA cfa, DOTBuilder2 dotBuilder, @Nullable CounterexampleInfo counterExample)
@@ -352,6 +388,7 @@ public class ReportGenerator {
     writer.write(",\n");
     dotBuilder.writeCfaInfo(writer);
     writer.write("\n}\n");
+    writer.write("window.cfaJson = cfaJson;\n");
   }
 
   private void insertArgJson(Writer writer) throws IOException {
@@ -378,6 +415,7 @@ public class ReportGenerator {
       writer.write("\n");
     }
     writer.write("}\n");
+    writer.write("window.argJson = argJson;\n");
   }
 
   private void insertCss(Writer writer) throws IOException {
@@ -593,6 +631,7 @@ public class ReportGenerator {
     writer.write("var sourceFiles = ");
     JSON.writeJSONString(sourceFiles, writer);
     writer.write(";\n");
+    writer.write("window.sourceFiles = sourceFiles;\n");
   }
 
   /** Build ARG data for all ARG states in the reached set. */
