+++ src/org/sosy_lab/cpachecker/core/counterexample/karma.conf.js
@@ -11,8 +11,12 @@
 
 // Karma configuration
 // Generated on Fri Jun 29 2018 17:07:20 GMT+0530 (India Standard Time)
-var fs = require('fs');
+const fs = require('fs');
+const path = require("path");
 const isDocker = require('is-docker')();
+const webpack = require('webpack');
+const webpackConfig = require('./webpack.config.js');
+const devFolder = "development_data/";
 
 // Replace scripts tag from HTML file and generate testReport.html for testing:
 // Remove everything from the line with REPORT_CSS to the line after REPORT_JS
@@ -25,39 +29,30 @@ fs.writeFileSync('testReport.html',
 module.exports = function (config) {
   config.set({
 
-    // base path that will be used to resolve all patterns (eg. files, exclude)
-    basePath: '',
-
     // frameworks to use
     // available frameworks: https://npmjs.org/browse/keyword/karma-adapter
-    frameworks: ['jasmine', 'detectBrowsers'],
+    frameworks: ['jasmine', 'jquery-3.4.0', 'detectBrowsers'],
 
     // list of files / patterns to load in the browser
     files: [
-      'https://cdnjs.cloudflare.com/ajax/libs/babel-core/5.8.34/browser-polyfill.js',
-      'https://www.sosy-lab.org/lib/angularjs/1.7.0/angular.min.js',
-      'https://cdnjs.cloudflare.com/ajax/libs/angular-mocks/1.7.2/angular-mocks.min.js',
-      'https://cdnjs.cloudflare.com/ajax/libs/angular-resource/1.7.2/angular-resource.min.js',
-      'https://www.sosy-lab.org/lib/jquery/3.3.1/jquery.min.js',
-      'https://www.sosy-lab.org/lib/jquery-ui/1.12.1/jquery-ui.min.js',
-      'https://cdn.jsdelivr.net/npm/jasmine-jquery@2.1.1/lib/jasmine-jquery.min.js',
-      'https://www.sosy-lab.org/lib/bootstrap/4.1.1/js/bootstrap.min.js',
-      'https://www.sosy-lab.org/lib/d3js/5.4.0/d3.min.js',
-      'https://www.sosy-lab.org/lib/dagre-d3/0.5.0/dagre-d3.min.js',
-      'https://www.sosy-lab.org/lib/datatables/1.10.18/datatables.min.js',
-      'https://www.sosy-lab.org/lib/popper.js/1.14.3/umd/popper.min.js',
-      'https://www.sosy-lab.org/lib/google-code-prettify/2018-04-29-453bd5f/prettify.js',
-      'testReport.html',
-      'report.js',
-      'test/*.js',
+      {pattern: 'report.js', watched: false},
+      {pattern: './node_modules/angular-mocks/ngMock.js', watched: false},
+      {pattern: './node_modules/jasmine-jquery/lib/jasmine-jquery.js', watched: false},
+      {pattern: 'testReport.html'},
+      {pattern: 'test/*.js', watched: false},
     ],
-
-    // list of files / patterns to exclude
-    exclude: [],
 
     // preprocess matching files before serving them to the browser
     // available preprocessors: https://npmjs.org/browse/keyword/karma-preprocessor
-    preprocessors: {},
+    preprocessors: {
+     'report.js': ['webpack'],
+     './node_modules/angular-mocks/ngMock.js': ['webpack'],
+     './node_modules/jasmine-jquery/lib/jasmine-jquery.js': ['webpack'],
+     'test/*.js': ['webpack'],
+   },
+
+    // list of files / patterns to exclude
+    exclude: ['test/initFunctionTest.js'],
 
     // test results reporter to use
     // possible values: 'dots', 'progress'
@@ -75,8 +70,10 @@ module.exports = function (config) {
       'karma-htmlfile-reporter',
       'karma-firefox-launcher',
       'karma-chrome-launcher',
+      'karma-detect-browsers',
       'karma-jasmine',
-      'karma-detect-browsers'
+      'karma-jquery',
+      'karma-webpack'
     ],
 
     detectBrowsers: {
@@ -124,6 +121,14 @@ module.exports = function (config) {
 
     // Concurrency level
     // how many browser should be started simultaneous
-    concurrency: Infinity
+    concurrency: Infinity,
+
+    webpack: {
+      mode: "development",
+      plugins: webpackConfig.plugins,
+      module: webpackConfig.module,
+      optimization: webpackConfig.optimization,
+      resolve: webpackConfig.resolve,
+    },
   })
 }
