+++ src/org/sosy_lab/cpachecker/core/counterexample/package.json
@@ -1,26 +1,56 @@
 {
-  "name": "counterexample",
-  "version": "1.0.0",
-  "description": "Report gererated by the CPAchecker",
-  "main": "report.js",
-  "scripts": {
-    "test": "karma start karma.conf.js",
-    "e2e-test": "protractor e2e-conf.js"
-  },
-  "author": "",
-  "license": "ISC",
-  "devDependencies": {
-    "babel-polyfill": "^6.26.0",
-    "karma": "^4.3.0",
-    "karma-chrome-launcher": "^3.1.0",
-    "karma-cli": "^2.0.0",
-    "karma-htmlfile-reporter": "^0.3.8",
-    "karma-jasmine": "^2.0.1",
-    "protractor-jasmine2-screenshot-reporter": "^0.5.0"
-  },
-  "dependencies": {
-    "karma-detect-browsers": "^2.3.3",
-    "karma-firefox-launcher": "^1.3.0",
-    "protractor": "^5.4.2"
-  }
+	"name": "counterexample",
+	"version": "1.0.0",
+	"description": "Report gererated by the CPAchecker",
+	"author": "",
+	"license": "ISC",
+	"private": true,
+	"scripts": {
+		"build": "webpack",
+		"start": "npm run build & webpack-dev-server --mode=development",
+		"test": "karma start karma.conf.js",
+		"e2e-test": "protractor e2e-conf.js"
+	},
+	"devDependencies": {
+		"babel-polyfill": "^6.26.0",
+		"karma": "^5.2.3",
+		"karma-chrome-launcher": "^3.1.0",
+		"karma-cli": "^2.0.0",
+		"karma-detect-browsers": "^2.3.3",
+		"karma-firefox-launcher": "^1.3.0",
+		"karma-htmlfile-reporter": "^0.3.8",
+		"karma-jasmine": "^2.0.1",
+		"karma-webpack": "^4.0.2",
+		"protractor": "^7.0.0",
+		"protractor-jasmine2-screenshot-reporter": "^0.5.0",
+		"webpack-dev-server": "^3.11.0"
+	},
+	"dependencies": {
+		"@babel/core": "^7.11.6",
+		"@fortawesome/fontawesome-free": "^5.15.1",
+		"@popperjs/core": "^2.5.3",
+		"angular": "^1.8.1",
+		"angular-mocks": "^1.8.2",
+		"bootstrap": "^4.5.2",
+		"browser-polyfill": "^3.20.2",
+		"code-prettify": "^0.1.0",
+		"css-loader": "^4.3.0",
+		"css-minimizer-webpack-plugin": "^1.1.5",
+		"d3": "^6.3.0",
+		"dagre-d3": "^0.6.4",
+		"datatables.net": "^1.10.22",
+		"datatables.net-dt": "^1.10.22",
+		"html-webpack-plugin": "^4.5.0",
+		"jasmine": "^3.6.3",
+		"jasmine-jquery": "^2.1.1",
+		"jquery": "^3.5.1",
+		"jquery-resizable-dom": "^0.35.0",
+		"karma-jquery": "^0.2.4",
+		"mini-css-extract-plugin": "^1.3.2",
+		"popper.js": "^1.16.1",
+		"tooltip": "^1.6.1",
+		"url-loader": "^4.1.1",
+		"webpack": "^4.44.2",
+		"webpack-cli": "^3.3.12"
+	}
 }
