+++ src/org/sosy_lab/cpachecker/core/counterexample/report.html
@@ -20,42 +20,12 @@ with considerably less effort -->
   <!-- METATAGS -->
   <title>CPAchecker Report</title>
   <!-- REPORT_CSS -->
-  <link rel="stylesheet" href="https://www.sosy-lab.org/lib/fontawesome-free/css/fontawesome-all.min.css" integrity="sha384-DNOHZ68U8hZfKXOrtjWvjxusGo9WQnrNx2sqG0tfsghAvtVlRW3tvkXWZh58N9jp"
-    crossorigin="anonymous">
-  <link rel="stylesheet" href="https://www.sosy-lab.org/lib/bootstrap/4.1.1/css/bootstrap.min.css" integrity="sha384-WskhaSGFgHYWDcbwN70/dfYBj47jz9qbsMId/iRN3ewGhXQFZCSftd1LZCfmhktB"
-    crossorigin="anonymous">
-  <script src="https://www.sosy-lab.org/lib/jquery/3.3.1/jquery.min.js" integrity="sha384-tsQFqpEReu7ZLhBV2VZlAu7zcOV+rXbYlF2cqB8txI/8aZajjp4Bqd+V6D5IgvKT"
-    crossorigin="anonymous"></script>
-  <link rel="stylesheet" href="https://www.sosy-lab.org/lib/jquery-ui/1.12.1/jquery-ui.min.css" integrity="sha384-VK/ia2DWrvtO05YDcbWI8WE3WciOH0RhfPNuRJGSa3dpAs5szXWQuCnPNv/yzpO4"
-    crossorigin="anonymous">
-  <script src="https://www.sosy-lab.org/lib/jquery-ui/1.12.1/jquery-ui.min.js" integrity="sha384-PtTRqvDhycIBU6x1wwIqnbDo8adeWIWP3AHmnrvccafo35E7oIvW7HPXn2YimvWu"
-    crossorigin="anonymous"></script>
-  <script src="https://www.sosy-lab.org/lib/popper.js/1.14.3/umd/popper.min.js" integrity="sha384-ZMP7rVo3mIykV+2+9J3UJ46jBk0WLaUAdn689aCwoqbBJiSnjAK/l8WvCWPIPm49"
-    crossorigin="anonymous"></script>
-  <script src="https://www.sosy-lab.org/lib/tooltip.js/1.2.0/tooltip.min.js" integrity="sha384-9MYfW4+cUynMIN6UiFuOu/5GAs2JEbbknCAyCvLPttuYxlLpocQpEkX210OoGn4p"
-    crossorigin="anonymous"></script>
-  <script src="https://www.sosy-lab.org/lib/angularjs/1.7.0/angular.min.js" integrity="sha384-OGKaWeggmIvCx/zchfRERRmxXYyFHhSRKBcld3xfy5N42zsVueW6rVyiL7E53ZGJ"
-    crossorigin="anonymous"></script>
-  <script src="https://www.sosy-lab.org/lib/d3js/5.4.0/d3.min.js" integrity="sha384-Ha//YQhNUuPzmPKgC8aYpurZcTiV3JhEvEhH5uGW0IRjBfEgDQxvo0aF1Gef+blI"
-    crossorigin="anonymous"></script>
-  <script src="https://www.sosy-lab.org/lib/dagre-d3/0.5.0/dagre-d3.min.js" integrity="sha384-0qz1K6Ihi4XT9AwqcOOC/ciBQ9ufdzKZUc8G1It6RZ+oLJSk41EhaZiLPVoBE0JT"
-    crossorigin="anonymous"></script>
-  <script src="https://www.sosy-lab.org/lib/google-code-prettify/2018-04-29-453bd5f/prettify.js" integrity="sha384-KlbExXRJYYILJYRAJR121vJXgBDx+puqf1yjznMcTt7C5njwQajOLFwk6pMafFnX"
-    crossorigin="anonymous"></script>
-  <link rel="stylesheet" href="https://www.sosy-lab.org/lib/google-code-prettify/2018-04-29-453bd5f/prettify.css" integrity="sha384-tGeLopS7aWCwgeqg+ah7c+iI19JU3teQPshyaQVybvOMRravAVLscwoeT4HPsfoW"
-    crossorigin="anonymous">
-  <script src="https://www.sosy-lab.org/lib/bootstrap/4.1.1/js/bootstrap.min.js" integrity="sha384-smHYKdLADwkXOn1EmN1qk/HfnUcbVRZyYmZ4qpPea6sjB/pTJ0euyQp0Mk8ck+5T"
-    crossorigin="anonymous"></script>
-  <link rel="stylesheet" href="https://www.sosy-lab.org/lib/datatables/1.10.18/datatables.min.css" integrity="sha384-t1jrh3H+gcVq93cU14/FvFkiDDLHAupR2uFWQFOIqqB0ZgNvd09e5yjT5Mc9mGPZ"
-    crossorigin="anonymous">
-  <script src="https://www.sosy-lab.org/lib/datatables/1.10.18/datatables.min.js" integrity="sha384-b6lz7vBoPNFWulOisoJlEziEWnbO6TbeulJvmrmuro+kXqlUgmn+sVfASUlqVcZo"
-    crossorigin="anonymous"></script>
   <script>
     <!-- REPORT_JS -->
   </script>
 </head>
 
-<body ng-controller="ReportController" id="report-controller" onload="init()">
+<body ng-controller="ReportController" id="report-controller" onload="window.init()">
   <header class="header">
     <a href="https://cpachecker.sosy-lab.org/">
       <img class="logo" ng-src="{{logo}}" />
